Adding a new feature: Target Query Validator.
This feature will check the consistency of the terms used by the target query against the ontology vocabulary. This addition extends the !MappingManagerPanel constructor in which it requires the active ontology used by Protégé. The validator then applies this ontology object to validate the target query. If the query is invalid, a warning message appears and users can't insert the mapping.

Currently, this feature is installed in the new mapping dialog. In the future, we are going to install it in the edit mapping panel.

BUG=42

Ticket URL=https://babbage.inf.unibz.it/trac/obdapublic/ticket/42